consolidate coesa core code, jited
